+++ stackslib/src/clarity_vm/tests/analysis_costs.rs
@@ -81,7 +81,7 @@ fn setup_tracked_cost_test(
     clarity_instance
         .begin_test_genesis_block(
             &StacksBlockId::sentinel(),
-            &StacksBlockId([0 as u8; 32]),
+            &StacksBlockId([0; 32]),
             &TEST_HEADER_DB,
             &burn_state_db,
         )
@@ -89,8 +89,8 @@ fn setup_tracked_cost_test(
 
     {
         let mut conn = clarity_instance.begin_block(
-            &StacksBlockId([0 as u8; 32]),
-            &StacksBlockId([1 as u8; 32]),
+            &StacksBlockId([0; 32]),
+            &StacksBlockId([1; 32]),
             &TEST_HEADER_DB,
             &burn_state_db,
         );
@@ -107,8 +107,8 @@ fn setup_tracked_cost_test(
 
     {
         let mut conn = clarity_instance.begin_block(
-            &StacksBlockId([1 as u8; 32]),
-            &StacksBlockId([2 as u8; 32]),
+            &StacksBlockId([1; 32]),
+            &StacksBlockId([2; 32]),
             &TEST_HEADER_DB,
             &burn_state_db,
         );
@@ -146,8 +146,8 @@ fn setup_tracked_cost_test(
 
     {
         let mut conn = clarity_instance.begin_block(
-            &StacksBlockId([2 as u8; 32]),
-            &StacksBlockId([3 as u8; 32]),
+            &StacksBlockId([2; 32]),
+            &StacksBlockId([3; 32]),
             &TEST_HEADER_DB,
             &burn_state_db,
         );
@@ -223,7 +223,7 @@ fn test_tracked_costs(
 
     {
         let mut conn = clarity_instance.begin_block(
-            &StacksBlockId([3 as u8; 32]),
+            &StacksBlockId([3; 32]),
             &StacksBlockId([4 + prog_id as u8; 32]),
             &TEST_HEADER_DB,
             &burn_state_db,
